[refactor/#269] 오늘의 루틴에 메시지 병합
